+++ src/organisms/feedCard/feedCard.styles.ts
